+++ packages/plugin-pha/src/typings.d.ts
@@ -1,0 +1,4 @@
+declare module '@ali/universal-mtop' {
+  const request: any;
+  export { request };
+}